more component cleanup
